// packages/app/src/urls.d.ts
export declare const URLS: {
    USER: {
        SIGNIN: string;
        SIGNUP: string;
        START_RESET_PASSWORD: string;
        RESET_PASSWORD: string;
    };
    TOKENS: {
        REFRESH_TOKEN: string;
        MAIL_INVITATION: string;
    };
    BET: {
        AVAILABLE: string;
        USER: string;
        CREATE: string;
    };
};